requirements.txt
flask
torch
pillow
opencv-python
ultralytics
gunicorn